requirements.txt
tensorflow
torch
dvc
notebook
numpy
pandas
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
transformers
datasets
matplotlib
seaborn
pandas
numpy
keras
opencv-python
scikit-learn
kaggle
python-dotenv
reusables
huggingface-hub
ipywidgets